Refactor functions into traits

+++ src/main/scala/com/translationdata/mixins/Palindrome.scala
@@ -1,6 +1,6 @@
 package com.translationdata.mixins
 trait Palindrome {
-  def isPalindrome(n:Int):Boolean = {
+  def isPalindrome(n:BigInt):Boolean = {
     val number = n.toString
     for ( leftPos <- 0 to number.length / 2 ) {
       val rightPos = number.length - leftPos - 1
